mysite/requirements.txt
Django==5.0.6
python-decouple==3.8
Requests==2.32.3
djangorestframework==3.15.2
